feat: move attestation server to separate command
