+++ cmd/swisstronikd/cmd/enclave.go
@@ -13,7 +13,7 @@ import (
 
 const flagShouldReset = "reset"
 
-// Cmd creates a CLI main command
+// EnclaveCmd creates a CLI main command
 func EnclaveCmd() *cobra.Command {
 	cmd := &cobra.Command{
 		Use:   "enclave",
@@ -24,7 +24,6 @@ func EnclaveCmd() *cobra.Command {
 	cmd.AddCommand(
 		EPIDRemoteAttestationCmd(),
 		DCAPRemoteAttestationCmd(),
-		StartAttestationServer(),
 		Status(),
 	)
 	return cmd
@@ -35,10 +34,10 @@ func EPIDRemoteAttestationCmd() *cobra.Command {
 	cmd := &cobra.Command{
 		Use:   "request-master-key-epid [bootstrap-node-address]",
 		Short: "Requests master key from bootstrap node using EPID",
-		Long:  `Initializes SGX enclave by passing process of EPID Remote Attestation agains bootstrap node. 
+		Long: `Initializes SGX enclave by passing process of EPID Remote Attestation agains bootstrap node. 
 		If remote attestation was successful, bootstrap node shares encrypted master key with this node. 
 		Process of Remote Attestation is performed over pure TCP protocol.`,
-		Args:  cobra.ExactArgs(1),
+		Args: cobra.ExactArgs(1),
 		RunE: func(cmd *cobra.Command, args []string) error {
 			clientCtx := client.GetClientContextFromCmd(cmd)
 			serverCtx := server.GetServerContextFromCmd(cmd)
@@ -72,10 +71,10 @@ func DCAPRemoteAttestationCmd() *cobra.Command {
 	cmd := &cobra.Command{
 		Use:   "request-master-key-dcap [bootstrap-node-address]",
 		Short: "Requests master key from bootstrap node using DCAP",
-		Long:  `Initializes SGX enclave by passing process of DCAP Remote Attestation agains bootstrap node. 
+		Long: `Initializes SGX enclave by passing process of DCAP Remote Attestation agains bootstrap node. 
 		If remote attestation was successful, bootstrap node shares encrypted master key with this node. 
 		Process of Remote Attestation is performed over pure TCP protocol.`,
-		Args:  cobra.ExactArgs(1),
+		Args: cobra.ExactArgs(1),
 		RunE: func(cmd *cobra.Command, args []string) error {
 			clientCtx := client.GetClientContextFromCmd(cmd)
 			serverCtx := server.GetServerContextFromCmd(cmd)
@@ -104,25 +103,7 @@ func DCAPRemoteAttestationCmd() *cobra.Command {
 	return cmd
 }
 
-// StartAttestationServer returns start-attestation-server cobra Command.
-func StartAttestationServer() *cobra.Command {
-	cmd := &cobra.Command{
-		Use:   "start-attestation-server [epid-address-with-port] [dcap-address-with-port]",
-		Short: "Starts attestation server",
-		Long:  "Start server for Intel SGX Remote Attestation to share master key with new nodes",
-		Args:  cobra.ExactArgs(2),
-		RunE: func(_ *cobra.Command, args []string) error {
-			if err := librustgo.StartAttestationServer(args[0], args[1]); err != nil {
-				return err
-			}
-			return server.WaitForQuitSignals()
-		},
-	}
-
-	return cmd
-}
-
-// Healthcheck checks if Intel SGX Enclave is accessible and if Intel SGX was properly configured
+// Status checks if Intel SGX Enclave is accessible and if Intel SGX was properly configured
 func Status() *cobra.Command {
 	cmd := &cobra.Command{
 		Use:   "status",
