+++ go-sgxvm/cmd/attestation/main.go
@@ -1,0 +1,9 @@
+package main
+
+import "github.com/SigmaGmbH/librustgo/cmd/attestation/cmd"
+
+func main() {
+	if err := cmd.RootCmd().Execute(); err != nil {
+		panic(err)
+	}
+}
