+++ go-sgxvm/go.mod
@@ -1,14 +1,12 @@
 module github.com/SigmaGmbH/librustgo
 
-go 1.18
+go 1.21
 
 require (
 	github.com/ethereum/go-ethereum v1.10.26
-	github.com/gorilla/mux v1.8.0
 	github.com/hashicorp/go-memdb v1.3.4
-	github.com/rs/cors v1.7.0
+	github.com/spf13/cobra v1.8.0
 	github.com/stretchr/testify v1.8.2
-	golang.org/x/net v0.0.0-20220607020251-c690dde0001d
 	google.golang.org/protobuf v1.30.0
 )
 
@@ -18,8 +16,10 @@ require (
 	github.com/decred/dcrd/dcrec/secp256k1/v4 v4.0.1 // indirect
 	github.com/hashicorp/go-immutable-radix v1.3.0 // indirect
 	github.com/hashicorp/golang-lru v0.5.5-0.20210104140557-80c98217689d // indirect
+	github.com/inconshreveable/mousetrap v1.1.0 // indirect
 	github.com/kr/pretty v0.1.0 // indirect
 	github.com/pmezard/go-difflib v1.0.0 // indirect
+	github.com/spf13/pflag v1.0.5 // indirect
 	golang.org/x/crypto v0.0.0-20210921155107-089bfa567519 // indirect
 	golang.org/x/sys v0.0.0-20220520151302-bc2c85ada10a // indirect
 	gopkg.in/check.v1 v1.0.0-20190902080502-41f04d3bba15 // indirect
